Move new utils to 'utils/search'
